ai-engineer-assignment/requirements.txt
# Core dependencies
langchain>=0.0.200
langchain-community>=0.0.200
langgraph>=0.0.40
streamlit>=1.24.0

# Data processing
pypdf>=3.0.0
requests>=2.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Configuration and validation
python-dotenv>=1.0.0
pydantic>=1.10.0
pydantic-settings>=2.0.0

# LLM providers
langchain-google-genai
google-generativeai

langsmith

# Optional dependencies (uncomment as needed)
# google-cloud-aiplatform>=1.30.0   # for Gemini (Vertex AI) integration
# langsmith>=0.2.0                  # for LangSmith integration
# openai>=1.0.0                     # for OpenAI integration

# Development dependencies (install with: pip install -e ".[dev]")
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
# mypy>=1.0.0
# pre-commit>=3.0.0
